makepkg: always honor --clean, not just if we were building something
This way, we will honor `--nobuild --clean` which on the surface might
not make sense, but this invocation is useful to e. g. run pkgver() and
clean up the source checkout (if for whatever reason we don't want it,
e. g. if the actual build will happen in a chroot).