repo-add: unconditionally create the database if it is missing
In cc8678ac (repo-add: do not recreate the database if nothing was changed),
repo-add was made to skip database write-out if there were no changes to
the database. However, this breaks the usage of repo-add to create a new
empty database: `repo-add /path/to/mydb/mydb.db.tar.xz`.

Bring back support for this use-case by always writing the database if
it is missing.

Fixes: cc8678ac461207257e7dd58eb57a476cf56580ec